requirements.txt
Django~=3.2.9
pip~=21.1.1
tornado~=6.1
pytz~=2021.3
sqlparse~=0.4.2
asgiref~=3.4.1
colorama~=0.4.4
Jinja2~=3.0.2
numpy~=1.21.3
setuptools~=56.0.0
djangorestframework~=3.12.4
Pygments~=2.10.0